+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Current
+
+### Changed
+
+- Allow the users to specify socket options and opt out of HTTP/1.1 or HTTP2.
+
 ## [0.18.8](https://github.com/CrowdHailer/Ace/tree/0.18.8) - 2019-04-29
 
 ### Fixed